Refactor: extract peft modules to peft.py & clean up vision_transformer.py
