+++ src/otx/backend/native/models/classification/hlabel_models/vit.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 import warnings
 from pathlib import Path
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Literal
 from urllib.parse import urlparse
 
 from torch import nn
@@ -71,14 +71,13 @@ class VisionTransformerHLabelCls(ForwardExplainMixInForViT, OTXHlabelClsModel):
         label_info: HLabelInfo,
         data_input_params: DataInputParams,
         model_name: str = "vit-tiny",
-        lora: bool = False,
+        peft: Literal["lora", "dora"] | None = None,
         optimizer: OptimizerCallable = DefaultOptimizerCallable,
         scheduler: LRSchedulerCallable | LRSchedulerListCallable = DefaultSchedulerCallable,
         metric: MetricCallable = HLabelClsMetricCallable,
         torch_compile: bool = False,
     ) -> None:
-        self.lora = lora
-
+        self.peft = peft
         super().__init__(
             label_info=label_info,
             data_input_params=data_input_params,
@@ -113,7 +112,6 @@ class VisionTransformerHLabelCls(ForwardExplainMixInForViT, OTXHlabelClsModel):
         vit_backbone = VisionTransformer(
             model_name=self.model_name,
             img_size=self.data_input_params.input_size,
-            peft=self.peft,
         )
         model = HLabelClassifier(
             backbone=vit_backbone,
