+++ src/otx/backend/native/models/classification/multilabel_models/vit.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 import warnings
 from pathlib import Path
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Literal
 from urllib.parse import urlparse
 
 from torch.hub import download_url_to_file
@@ -60,13 +60,13 @@ class VisionTransformerMultilabelCls(ForwardExplainMixInForViT, OTXMultilabelCls
         label_info: LabelInfoTypes,
         data_input_params: DataInputParams,
         model_name: str = "vit-tiny",
-        lora: bool = False,
+        peft: Literal["lora", "dora"] | None = None,
         optimizer: OptimizerCallable = DefaultOptimizerCallable,
         scheduler: LRSchedulerCallable | LRSchedulerListCallable = DefaultSchedulerCallable,
         metric: MetricCallable = MultiLabelClsMetricCallable,
         torch_compile: bool = False,
     ) -> None:
-        self.lora = lora
+        self.peft = peft
 
         super().__init__(
             label_info=label_info,
@@ -96,7 +96,6 @@ class VisionTransformerMultilabelCls(ForwardExplainMixInForViT, OTXMultilabelCls
         vit_backbone = VisionTransformer(
             model_name=self.model_name,
             img_size=self.data_input_params.input_size,
-            peft=self.peft,
         )
         model = ImageClassifier(
             backbone=vit_backbone,
